+++ typescript/infra/src/config/gas-oracle.ts
@@ -213,26 +213,16 @@ function getMinUsdCost(local: ChainName, remote: ChainName): number {
   }
 
   const remoteMinCostOverrides: ChainMap<number> = {
+    // mitosis
+    mitosis: 0.1,
+
+    // For all SVM chains, min cost is 0.50 USD to cover rent needs
     // For Ethereum L2s, we need to account for the L1 DA costs that
     // aren't accounted for directly in the gas price.
-    arbitrum: 0.5,
     ancient8: 0.5,
     blast: 0.5,
-    bob: 0.5,
-    linea: 0.5,
     mantapacific: 0.5,
-    mantle: 0.5,
     polygonzkevm: 0.5,
-
-    // op stack chains
-    base: 0.2,
-    fraxtal: 0.2,
-    lisk: 0.2,
-    mode: 0.2,
-    optimism: 0.2,
-    soneium: 0.2,
-    superseed: 0.2,
-    unichain: 0.2,
 
     // Scroll is more expensive than the rest due to higher L1 fees
     scroll: 1.5,