+++ typescript/utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @hyperlane-xyz/utils
 
+## 18.2.0
+
+## 18.1.0
+
 ## 18.0.0
 
 ### Minor Changes